+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Changed some equalities with FhirTime
 * Added partial timezones for FhirDateTime, FhirInstant
+* Added precision comparison for DateTimePrecision
 
 ## [0.12.0]